+++ dyn_array.h
@@ -17,7 +17,7 @@ extern void dyn_array_delete_traders(dyn_array* dyn, int index);
 extern void dyn_array_free_products(dyn_array* dyn);
 extern trader* dyn_array_get_trader(dyn_array* dyn, pid_t pid);
 extern product* dyn_array_get_product(dyn_array* dyn, char* name);
-price_entry* dyn_array_get_price_entry(dyn_array* dyn, int value);
+price_entry* dyn_array_get_price_entry(dyn_array* dyn, long value);
 void dyn_array_add_price(dyn_array* dyn, void* value);
 
 #endif
