+++ functions.h
@@ -15,9 +15,9 @@ extern bool validate_buysell(char* command);
 extern void tell_other_traders(exchange* pexchange, int id, char* message);
 extern int read_command(int fd, char* buffer);
 extern void print_report(exchange* pexchange);
-extern order* store_product(exchange* pexchange, trader* source, command command_type, int order_id, char* product_name, int qty, int price);
-void match_order(exchange* pexchange, command command_type, char* product_name, int price, order* new_order, trader* source);
-void amend_order(exchange* pexchange, order* to_amend, int qty, int price);
+extern order* store_product(exchange* pexchange, trader* source, command command_type, int order_id, char* product_name, long qty, long price);
+void match_order(exchange* pexchange, command command_type, char* product_name, long price, order* new_order, trader* source);
+void amend_order(exchange* pexchange, order* to_amend, long qty, long price);
 void cancel_order(exchange* pexchange, order* found);
 
 #endif
